fix: lower the hero's height budget to account for the status label
HeroWaveformView.maximumHeight (80) and Preferences's default height (64)
were both sized against a margin measured with statusLabel hidden, but
refreshForCurrentClip() shows it ("Separating in the background…") for as
long as a clip is still separating — a common state, not an edge case. With
it visible, the maximum-height case overflowed WindowSizing.minimum.height
by 15pt and the default case was down to ~1pt of margin.

Add a testing seam (statusLabelForTesting) and two WindowSizingTests that
measure the deck with the status label visible, at both the default and
maximum hero heights. Measured: maximum needed 588pt (was 615pt) and default
needed 580pt (was 599pt) against the 600pt floor — both now with real margin.
Lower HeroWaveformView.maximumHeight to 53 and the registered default height
to 45, and update the doc comments that quoted the old figures.

+++ Sources/MinusOne/Preferences.swift
@@ -38,7 +38,11 @@ final class Preferences {
             Key.appearance: AppAppearance.system.rawValue,
             Key.recordingSource: RecordingSource.systemAudio.storedValue,
             Key.heroWaveformEnabled: true,
-            Key.heroWaveformHeight: Double(64)
+            // 45, not the original 64: `WindowSizingTests
+            // .testTheDeckFitsTheMinimumWindowHeightAtDefaultHeroHeightWithStatusVisible` measures the
+            // deck with `statusLabel` visible (a clip still separating — a common state, not an edge
+            // case) and 64 left only ~1pt of margin at `WindowSizing.minimum.height`. 45 leaves ~20pt.
+            Key.heroWaveformHeight: Double(45)
         ])
     }
